wait for handler to register
