+++ pkg/handler/handler.go
@@ -1,0 +1,126 @@
+// Copyright 2023 LiveKit, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package handler
+
+import (
+	"context"
+	"time"
+
+	"github.com/frostbyte73/core"
+	"google.golang.org/grpc"
+
+	"github.com/livekit/egress/pkg/config"
+	"github.com/livekit/egress/pkg/errors"
+	"github.com/livekit/egress/pkg/ipc"
+	"github.com/livekit/egress/pkg/pipeline"
+	"github.com/livekit/protocol/livekit"
+	"github.com/livekit/protocol/rpc"
+	"github.com/livekit/protocol/tracer"
+	"github.com/livekit/psrpc"
+)
+
+type Handler struct {
+	ipc.UnimplementedEgressHandlerServer
+
+	conf             *config.PipelineConfig
+	pipeline         *pipeline.Controller
+	rpcServer        rpc.EgressHandlerServer
+	ipcHandlerServer *grpc.Server
+	ipcServiceClient ipc.EgressServiceClient
+	ioClient         rpc.IOInfoClient
+	kill             core.Fuse
+}
+
+func NewHandler(conf *config.PipelineConfig, bus psrpc.MessageBus, ioClient rpc.IOInfoClient) (*Handler, error) {
+	ipcClient, err := ipc.NewServiceClient()
+	if err != nil {
+		return nil, err
+	}
+
+	h := &Handler{
+		conf:             conf,
+		ioClient:         ioClient,
+		ipcHandlerServer: grpc.NewServer(),
+		ipcServiceClient: ipcClient,
+		kill:             core.NewFuse(),
+	}
+
+	ipc.RegisterEgressHandlerServer(h.ipcHandlerServer, h)
+	if err = ipc.StartHandlerListener(h.ipcHandlerServer, conf.TmpDir); err != nil {
+		return nil, err
+	}
+
+	rpcServer, err := rpc.NewEgressHandlerServer(h, bus)
+	if err != nil {
+		return nil, errors.Fatal(err)
+	}
+	if err = rpcServer.RegisterUpdateStreamTopic(conf.Info.EgressId); err != nil {
+		return nil, errors.Fatal(err)
+	}
+	if err = rpcServer.RegisterStopEgressTopic(conf.Info.EgressId); err != nil {
+		return nil, errors.Fatal(err)
+	}
+	h.rpcServer = rpcServer
+
+	ctx := context.Background()
+	_, _ = h.ipcServiceClient.HandlerReady(ctx, &ipc.HandlerReadyRequest{EgressId: conf.Info.EgressId})
+
+	h.pipeline, err = pipeline.New(ctx, conf, h.ioClient)
+	if err != nil {
+		if !errors.IsFatal(err) {
+			// user error, send update
+			now := time.Now().UnixNano()
+			conf.Info.UpdatedAt = now
+			conf.Info.EndedAt = now
+			conf.Info.Status = livekit.EgressStatus_EGRESS_FAILED
+			conf.Info.Error = err.Error()
+			_, _ = h.ioClient.UpdateEgress(ctx, conf.Info)
+		}
+		return nil, err
+	}
+
+	return h, nil
+}
+
+func (h *Handler) Run() error {
+	ctx, span := tracer.Start(context.Background(), "Handler.Run")
+	defer span.End()
+
+	// start egress
+	result := make(chan *livekit.EgressInfo, 1)
+	go func() {
+		result <- h.pipeline.Run(ctx)
+	}()
+
+	kill := h.kill.Watch()
+	for {
+		select {
+		case <-kill:
+			// kill signal received
+			h.pipeline.SendEOS(ctx)
+
+		case res := <-result:
+			// recording finished
+			_, _ = h.ioClient.UpdateEgress(ctx, res)
+			h.rpcServer.Shutdown()
+			h.ipcHandlerServer.Stop()
+			return nil
+		}
+	}
+}
+
+func (h *Handler) Kill() {
+	h.kill.Break()
+}
