+++ pkg/service/process.go
@@ -16,41 +16,29 @@ package service
 
 import (
 	"context"
-	"net"
-	"os"
 	"os/exec"
-	"path"
 	"strings"
-	"syscall"
-	"time"
 
 	"github.com/frostbyte73/core"
 	dto "github.com/prometheus/client_model/go"
 	"github.com/prometheus/common/expfmt"
 	"golang.org/x/exp/maps"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-	"google.golang.org/protobuf/encoding/protojson"
-	"gopkg.in/yaml.v3"
 
-	"github.com/livekit/egress/pkg/config"
-	"github.com/livekit/egress/pkg/errors"
 	"github.com/livekit/egress/pkg/ipc"
 	"github.com/livekit/protocol/livekit"
 	"github.com/livekit/protocol/logger"
 	"github.com/livekit/protocol/rpc"
-	"github.com/livekit/protocol/tracer"
-	"github.com/livekit/protocol/utils"
 )
 
 type Process struct {
-	ctx        context.Context
-	handlerID  string
-	req        *rpc.StartEgressRequest
-	info       *livekit.EgressInfo
-	cmd        *exec.Cmd
-	grpcClient ipc.EgressHandlerClient
-	closed     core.Fuse
+	ctx              context.Context
+	handlerID        string
+	req              *rpc.StartEgressRequest
+	info             *livekit.EgressInfo
+	cmd              *exec.Cmd
+	ipcHandlerClient ipc.EgressHandlerClient
+	ready            chan struct{}
+	closed           core.Fuse
 }
 
 func NewProcess(
@@ -61,142 +49,29 @@ func NewProcess(
 	cmd *exec.Cmd,
 	tmpDir string,
 ) (*Process, error) {
-	p := &Process{
-		ctx:       ctx,
-		handlerID: handlerID,
-		req:       req,
-		info:      info,
-		cmd:       cmd,
-		closed:    core.NewFuse(),
-	}
-
-	socketAddr := getSocketAddress(tmpDir)
-	conn, err := grpc.Dial(socketAddr,
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithContextDialer(func(_ context.Context, addr string) (net.Conn, error) {
-			return net.Dial(network, addr)
-		}),
-	)
+	ipcClient, err := ipc.NewHandlerClient(tmpDir)
 	if err != nil {
-		logger.Errorw("could not dial grpc handler", err)
 		return nil, err
 	}
-	p.grpcClient = ipc.NewEgressHandlerClient(conn)
+
+	p := &Process{
+		ctx:              ctx,
+		handlerID:        handlerID,
+		req:              req,
+		info:             info,
+		cmd:              cmd,
+		ipcHandlerClient: ipcClient,
+		ready:            make(chan struct{}),
+		closed:           core.NewFuse(),
+	}
 
 	return p, nil
-}
-
-func (s *Service) launchHandler(req *rpc.StartEgressRequest, info *livekit.EgressInfo) error {
-	_, span := tracer.Start(context.Background(), "Service.launchHandler")
-	defer span.End()
-
-	handlerID := utils.NewGuid("EGH_")
-	p := &config.PipelineConfig{
-		BaseConfig: s.conf.BaseConfig,
-		HandlerID:  handlerID,
-		TmpDir:     path.Join(os.TempDir(), handlerID),
-	}
-
-	confString, err := yaml.Marshal(p)
-	if err != nil {
-		span.RecordError(err)
-		logger.Errorw("could not marshal config", err)
-		return err
-	}
-
-	reqString, err := protojson.Marshal(req)
-	if err != nil {
-		span.RecordError(err)
-		logger.Errorw("could not marshal request", err)
-		return err
-	}
-
-	cmd := exec.Command("egress",
-		"run-handler",
-		"--config", string(confString),
-		"--request", string(reqString),
-	)
-	cmd.Dir = "/"
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	if err = cmd.Start(); err != nil {
-		span.RecordError(err)
-		logger.Errorw("could not launch process", err)
-		return err
-	}
-
-	s.EgressStarted(req)
-
-	h, err := NewProcess(context.Background(), handlerID, req, info, cmd, p.TmpDir)
-	if err != nil {
-		span.RecordError(err)
-		return err
-	}
-
-	s.AddHandler(req.EgressId, h)
-	return nil
-}
-
-func (s *Service) AddHandler(egressID string, h *Process) {
-	s.mu.Lock()
-	s.activeHandlers[egressID] = h
-	s.mu.Unlock()
-
-	go s.awaitCleanup(h)
-}
-
-func (s *Service) awaitCleanup(p *Process) {
-	if err := p.cmd.Wait(); err != nil {
-		now := time.Now().UnixNano()
-		p.info.UpdatedAt = now
-		p.info.EndedAt = now
-		p.info.Status = livekit.EgressStatus_EGRESS_FAILED
-		p.info.Error = "internal error"
-		_, _ = s.ioClient.UpdateEgress(p.ctx, p.info)
-		s.Stop(false)
-	}
-
-	s.EgressEnded(p.req)
-	p.closed.Break()
-
-	s.mu.Lock()
-	delete(s.activeHandlers, p.req.EgressId)
-	s.mu.Unlock()
-}
-
-func (s *Service) getGRPCClient(egressID string) (ipc.EgressHandlerClient, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	h, ok := s.activeHandlers[egressID]
-	if !ok {
-		return nil, errors.ErrEgressNotFound
-	}
-	return h.grpcClient, nil
-}
-
-func (s *Service) KillAll() {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	for _, h := range s.activeHandlers {
-		if !h.closed.IsBroken() {
-			if err := h.cmd.Process.Signal(syscall.SIGINT); err != nil {
-				logger.Errorw("failed to kill Process", err, "egressID", h.req.EgressId)
-			}
-		}
-	}
-}
-
-func getSocketAddress(handlerTmpDir string) string {
-	return path.Join(handlerTmpDir, "service_rpc.sock")
 }
 
 // Gather implements the prometheus.Gatherer interface on server-side to allow aggregation of handler metrics
 func (p *Process) Gather() ([]*dto.MetricFamily, error) {
 	// Get the metrics from the handler via IPC
-	metricsResponse, err := p.grpcClient.GetMetrics(context.Background(), &ipc.MetricsRequest{})
+	metricsResponse, err := p.ipcHandlerClient.GetMetrics(context.Background(), &ipc.MetricsRequest{})
 	if err != nil {
 		logger.Warnw("Error obtaining metrics from handler, skipping", err, "egress_id", p.req.EgressId)
 		return make([]*dto.MetricFamily, 0), nil // don't return an error, just skip this handler
@@ -210,15 +85,16 @@ func (p *Process) Gather() ([]*dto.MetricFamily, error) {
 	}
 
 	// Add an egress_id label to every metric all the families, if it doesn't already have one
-	applyDefaultLabel(p, families)
+	applyDefaultLabel(p.info.EgressId, families)
 
 	return maps.Values(families), nil
 }
 
-func applyDefaultLabel(p *Process, families map[string]*dto.MetricFamily) {
+func applyDefaultLabel(egressID string, families map[string]*dto.MetricFamily) {
+	egressIDLabel := "egress_id"
 	egressLabelPair := &dto.LabelPair{
-		Name:  StringPtr("egress_id"),
-		Value: &p.req.EgressId,
+		Name:  &egressIDLabel,
+		Value: &egressID,
 	}
 	for _, family := range families {
 		for _, metric := range family.Metric {
@@ -238,5 +114,3 @@ func applyDefaultLabel(p *Process, families map[string]*dto.MetricFamily) {
 		}
 	}
 }
-
-func StringPtr(v string) *string { return &v }
