+++ pkg/service/service.go
@@ -15,37 +15,38 @@
 package service
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
+	"io/fs"
 	"net"
 	"net/http"
 	"sync"
+	"syscall"
 	"time"
 
 	"github.com/frostbyte73/core"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
-	dto "github.com/prometheus/client_model/go"
+	"google.golang.org/grpc"
 
 	"github.com/livekit/egress/pkg/config"
+	"github.com/livekit/egress/pkg/ipc"
 	"github.com/livekit/egress/pkg/stats"
 	"github.com/livekit/egress/version"
-	"github.com/livekit/protocol/egress"
-	"github.com/livekit/protocol/livekit"
 	"github.com/livekit/protocol/logger"
 	"github.com/livekit/protocol/rpc"
-	"github.com/livekit/protocol/tracer"
 )
 
 const shutdownTimer = time.Second * 30
 
 type Service struct {
-	conf        *config.ServiceConfig
-	psrpcServer rpc.EgressInternalServer
-	ioClient    rpc.IOInfoClient
-	promServer  *http.Server
+	ipc.UnimplementedEgressServiceServer
+
 	*stats.Monitor
+
+	conf             *config.ServiceConfig
+	psrpcServer      rpc.EgressInternalServer
+	ipcServiceServer *grpc.Server
+	ioClient         rpc.IOInfoClient
+	promServer       *http.Server
 
 	mu             sync.RWMutex
 	activeHandlers map[string]*Process
@@ -54,14 +55,18 @@ type Service struct {
 }
 
 func NewService(conf *config.ServiceConfig, ioClient rpc.IOInfoClient) (*Service, error) {
-	monitor := stats.NewMonitor(conf)
-
 	s := &Service{
-		conf:           conf,
-		ioClient:       ioClient,
-		Monitor:        monitor,
-		shutdown:       core.NewFuse(),
-		activeHandlers: make(map[string]*Process),
+		Monitor:          stats.NewMonitor(conf),
+		conf:             conf,
+		ipcServiceServer: grpc.NewServer(),
+		ioClient:         ioClient,
+		shutdown:         core.NewFuse(),
+		activeHandlers:   make(map[string]*Process),
+	}
+
+	ipc.RegisterEgressServiceServer(s.ipcServiceServer, s)
+	if err := ipc.StartServiceListener(s.ipcServiceServer); err != nil {
+		return nil, err
 	}
 
 	if conf.PrometheusPort > 0 {
@@ -96,6 +101,26 @@ func (s *Service) Register(psrpcServer rpc.EgressInternalServer) {
 	s.psrpcServer = psrpcServer
 }
 
+func (s *Service) StartTemplatesServer(fs fs.FS) error {
+	if s.conf.TemplatePort == 0 {
+		logger.Debugw("templates server disabled")
+		return nil
+	}
+
+	h := http.FileServer(http.FS(fs))
+
+	mux := http.NewServeMux()
+	mux.Handle("/", h)
+
+	go func() {
+		addr := fmt.Sprintf("localhost:%d", s.conf.TemplatePort)
+		logger.Debugw(fmt.Sprintf("starting template server on address %s", addr))
+		_ = http.ListenAndServe(addr, mux)
+	}()
+
+	return nil
+}
+
 func (s *Service) Run() error {
 	logger.Debugw("starting service", "version", version.Version)
 
@@ -118,80 +143,6 @@ func (s *Service) Reset() {
 	s.shutdown = core.NewFuse()
 }
 
-func (s *Service) StartEgress(ctx context.Context, req *rpc.StartEgressRequest) (*livekit.EgressInfo, error) {
-	ctx, span := tracer.Start(ctx, "Service.StartEgress")
-	defer span.End()
-
-	if err := s.AcceptRequest(req); err != nil {
-		return nil, err
-	}
-
-	logger.Infow("request received", "egressID", req.EgressId)
-
-	p, err := config.GetValidatedPipelineConfig(s.conf, req)
-	if err != nil {
-		s.EgressAborted(req)
-		return nil, err
-	}
-
-	_, err = s.ioClient.CreateEgress(ctx, p.Info)
-	if err != nil {
-		s.EgressAborted(req)
-		return nil, err
-	}
-
-	requestType, outputType := egress.GetTypes(p.Info.Request)
-	logger.Infow("request validated",
-		"egressID", req.EgressId,
-		"requestType", requestType,
-		"outputType", outputType,
-		"room", p.Info.RoomName,
-		"request", p.Info.Request,
-	)
-
-	err = s.launchHandler(req, p.Info)
-	if err != nil {
-		s.EgressAborted(req)
-		return nil, err
-	}
-
-	return p.Info, nil
-}
-
-func (s *Service) StartEgressAffinity(ctx context.Context, req *rpc.StartEgressRequest) float32 {
-	if !s.CanAcceptRequest(req) {
-		// cannot accept
-		return -1
-	}
-
-	if s.GetRequestCount() == 0 {
-		// group multiple track and track composite requests.
-		// if this instance is idle and another is already handling some, the request will go to that server.
-		// this avoids having many instances with one track request each, taking availability from room composite.
-		return 0.5
-	} else {
-		// already handling a request and has available cpu
-		return 1
-	}
-}
-
-func (s *Service) ListActiveEgress(ctx context.Context, _ *rpc.ListActiveEgressRequest) (*rpc.ListActiveEgressResponse, error) {
-	ctx, span := tracer.Start(ctx, "Service.ListActiveEgress")
-	defer span.End()
-
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	egressIDs := make([]string, 0, len(s.activeHandlers))
-	for egressID := range s.activeHandlers {
-		egressIDs = append(egressIDs, egressID)
-	}
-
-	return &rpc.ListActiveEgressResponse{
-		EgressIds: egressIDs,
-	}, nil
-}
-
 func (s *Service) Status() ([]byte, error) {
 	info := map[string]interface{}{
 		"CpuLoad": s.GetCPULoad(),
@@ -206,39 +157,6 @@ func (s *Service) Status() ([]byte, error) {
 	return json.Marshal(info)
 }
 
-func (s *Service) promIsIdle() float64 {
-	if s.GetRequestCount() == 0 {
-		return 1
-	}
-	return 0
-}
-
-func (s *Service) promCanAcceptRequest() float64 {
-	if s.CanAcceptRequest(&rpc.StartEgressRequest{
-		Request: &rpc.StartEgressRequest_RoomComposite{
-			RoomComposite: &livekit.RoomCompositeEgressRequest{},
-		},
-	}) {
-		return 1
-	}
-	return 0
-}
-
-func (s *Service) promGetEgressIDs(pUsage map[int]float64) map[string]float64 {
-	s.mu.RLock()
-	pids := make(map[int]string)
-	for _, h := range s.activeHandlers {
-		pids[h.cmd.Process.Pid] = h.req.EgressId
-	}
-	s.mu.RUnlock()
-
-	eUsage := make(map[string]float64)
-	for pid, egressID := range pids {
-		eUsage[egressID] = pUsage[pid]
-	}
-	return eUsage
-}
-
 func (s *Service) Stop(kill bool) {
 	s.shutdown.Once(func() {
 		s.psrpcServer.DeregisterStartEgressTopic(s.conf.ClusterID)
@@ -248,35 +166,23 @@ func (s *Service) Stop(kill bool) {
 	}
 }
 
+func (s *Service) KillAll() {
+	s.mu.RLock()
+	defer s.mu.RUnlock()
+
+	for _, h := range s.activeHandlers {
+		if !h.closed.IsBroken() {
+			if err := h.cmd.Process.Signal(syscall.SIGINT); err != nil {
+				logger.Errorw("failed to kill Process", err, "egressID", h.req.EgressId)
+			}
+		}
+	}
+}
+
 func (s *Service) Close() {
 	for s.GetRequestCount() > 0 {
 		time.Sleep(shutdownTimer)
 	}
 	logger.Infow("closing server")
 	s.psrpcServer.Shutdown()
-}
-
-func (s *Service) CreateGatherer() prometheus.Gatherer {
-	return prometheus.GathererFunc(func() ([]*dto.MetricFamily, error) {
-		_, span := tracer.Start(context.Background(), "Service.GathererOfHandlerMetrics")
-		defer span.End()
-
-		s.mu.RLock()
-		defer s.mu.RUnlock()
-
-		gatherers := prometheus.Gatherers{}
-		// Include the default repo
-		gatherers = append(gatherers, prometheus.DefaultGatherer)
-		// add all the active handlers as sources
-		for _, v := range s.activeHandlers {
-			gatherers = append(gatherers, v)
-		}
-		return gatherers.Gather()
-	})
-}
-
-func (s *Service) PromHandler() http.Handler {
-	return promhttp.InstrumentMetricHandler(
-		prometheus.DefaultRegisterer, promhttp.HandlerFor(s.CreateGatherer(), promhttp.HandlerOpts{}),
-	)
 }
